+++ packages/components/src/components/formfieldwrapper/formfieldwrapper.types.ts
@@ -1,5 +1,5 @@
 import type { ValueOf } from '../../utils/types';
-import { VALIDATION } from './labelandhelper.constants';
+import { VALIDATION } from './formfieldwrapper.constants';
 
 type ValidationType = ValueOf<typeof VALIDATION>;
