+++ packages/components/src/components/formfieldwrapper/index.ts
@@ -1,0 +1,14 @@
+import FormfieldWrapper from './formfieldwrapper.component';
+import { TAG_NAME } from './formfieldwrapper.constants';
+import '../icon';
+import '../text';
+
+FormfieldWrapper.register(TAG_NAME);
+
+declare global {
+  interface HTMLElementTagNameMap {
+    ['mdc-formfieldwrapper']: FormfieldWrapper;
+  }
+}
+
+export default FormfieldWrapper;
